fix: vue3 pwa

+++ src/util/file/outputFileContents.js
@@ -9,16 +9,25 @@ export default function outputFileContents(appName, fileMap) {
   try {
     fileMap.forEach((files) => {
       const file = files[0];
-      const output = file.isJson
-        ? JSON.stringify(file.content, null, 2)
-        : file.content
+
+      if (file.isFile) {
+        fse.outputFileSync(
+          path.join(process.cwd(), appName, file.filename),
+          file.content,
+        );
+      } else {
+        const output = file.isJson
+          ? JSON.stringify(file.content, null, 2)
+          : file.content
             .replace(RegExp('(//<|<!)---=\\w+\\|\\|([^>]*)--->', 'g'), '$2')
             .replace(RegExp('(\\n+)( *)(//<|<!)(---\\+\\w+--->)', 'g'), '');
 
-      fse.outputFileSync(
-        path.join(process.cwd(), appName, file.filename),
-        composer.runOutputFilters(output, file.filename)
-      );
+
+        fse.outputFileSync(
+          path.join(process.cwd(), appName, file.filename),
+          composer.runOutputFilters(output, file.filename)
+        );
+      }
     });
     createFileSpinner.succeed(`${appName} project created`);
   } catch (e) {
